feat: 주문 생성 시 데이터 플랫폼 이벤트 리스너 구현

+++ ecommerce/src/main/java/com/ecommerce/domain/order/event/OrderCreatedEvent.java
@@ -1,13 +1,15 @@
 package com.ecommerce.domain.order.event;
 
 
+import com.ecommerce.domain.order.OrderItem;
+
 import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderCreatedEvent(
         Long orderId,
         Long userId,
-        BigDecimal pointBalance,
+        Long pointBalance,
         List<Long> cartItemIds
 ) {
 }
